+++ modules/utils/index.js
@@ -1,6 +1,7 @@
 'use strict';
 export {queryBuilderFormat} from './queryBuilderFormat'
 export {mongodbFormat} from './mongodbFormat'
+export {sqlFormat} from './sqlFormat'
 export {queryString} from './queryString'
 export {getTree, loadTree, checkTree} from './treeUtils'
 export {validateTree} from './validation'