+++ modules/utils/sqlFormat.js
@@ -5,8 +5,34 @@ import {defaultValue} from "./stuff";
 import {defaultConjunction} from './defaultUtils';
 import {settings as defaultSettings} from '../config/default';
 import {Map} from 'immutable';
+let SqlString = require('sqlstring');
 
-export const queryString = (item, config, isForDisplay = false) => {
+
+SqlString.trim = (val) => {
+    if (val.charAt(0) == "'")
+        return val.substring(1, val.length-1);
+    else
+        return val;
+};
+
+SqlString.escapeLike = (val) => {
+    // normal escape
+    let res = SqlString.escape(val);
+    // unwrap ''
+    res = SqlString.trim(res);
+    // escape % and _
+    res = res.replace(/[%_]/g, '\\$&');
+    // wrap with % for LIKE
+    res = "%" + res + "%";
+    // wrap ''
+    res = "'" + res + "'";
+    return res;
+};
+
+export {SqlString};
+
+
+export const sqlFormat = (item, config) => {
     const type = item.get('type');
     const properties = item.get('properties') || new Map();
     const children = item.get('children1');
@@ -14,7 +40,7 @@ export const queryString = (item, config, isForDisplay = false) => {
     if (type === 'group' && children && children.size) {
         const not = properties.get('not');
         const list = children
-            .map((currentChild) => queryString(currentChild, config, isForDisplay))
+            .map((currentChild) => sqlFormat(currentChild, config))
             .filter((currentChild) => typeof currentChild !== 'undefined');
         if (!list.size)
             return undefined;
@@ -24,7 +50,7 @@ export const queryString = (item, config, isForDisplay = false) => {
             conjunction = defaultConjunction(config);
         const conjunctionDefinition = config.conjunctions[conjunction];
 
-        return conjunctionDefinition.formatConj(list, conjunction, not, isForDisplay);
+        return conjunctionDefinition.sqlFormatConj(list, conjunction, not);
     } else if (type === 'rule') {
         let field = properties.get('field');
         const operator = properties.get('operator');
@@ -65,17 +91,22 @@ export const queryString = (item, config, isForDisplay = false) => {
                     const fieldFullLabel = fieldPartsLabels ? fieldPartsLabels.join(fieldSeparatorDisplay) : null;
                     const fieldLabel2 = rightFieldDefinition.label2 || fieldFullLabel;
                     const formatField = config.settings.formatField || defaultSettings.formatField;
-                    formattedField = formatField(rightField, fieldParts, fieldLabel2, rightFieldDefinition, config, isForDisplay);
+                    let rightFieldName = rightField;
+                    if (rightFieldDefinition.tableName) {
+                        const fieldPartsCopy = [...fieldParts];
+                        fieldPartsCopy[0] = rightFieldDefinition.tableName;
+                        rightFieldName = fieldPartsCopy.join(fieldSeparator);
+                    }
+                    formattedField = formatField(rightFieldName, fieldParts, fieldLabel2, rightFieldDefinition, config);
                 }
                 ret = formattedField;
             } else {
-                if (typeof fieldWidgetDefinition.formatValue === 'function') {
-                    const fn = fieldWidgetDefinition.formatValue;
+                if (typeof fieldWidgetDefinition.sqlFormatValue === 'function') {
+                    const fn = fieldWidgetDefinition.sqlFormatValue;
                     const args = [
                         currentValue,
                         pick(fieldDefinition, ['fieldSettings', 'listValues']),
                         omit(fieldWidgetDefinition, ['formatValue', 'mongoFormatValue', 'sqlFormatValue']), //useful options: valueFormat for date/time
-                        isForDisplay
                     ];
                     if (valueSrc == 'field') {
                         const valFieldDefinition = getFieldConfig(currentValue, config) || {}; 
@@ -87,7 +118,7 @@ export const queryString = (item, config, isForDisplay = false) => {
                     }
                     ret = fn(...args);
                 } else {
-                    ret = currentValue;
+                    ret = SqlString.escape(currentValue);
                 }
             }
             valueSrcs.push(valueSrc);
@@ -100,18 +131,30 @@ export const queryString = (item, config, isForDisplay = false) => {
 
         //find fn to format expr
         let isRev = false;
-        let fn = operatorDefinition.formatOp;
+        let fn = operatorDefinition.sqlFormatOp;
         if (!fn && reversedOp) {
-            fn = revOperatorDefinition.formatOp;
+            fn = revOperatorDefinition.sqlFormatOp;
             if (fn) {
                 isRev = true;
             }
         }
-        if (!fn && cardinality == 1) {
-            let _operator = operatorDefinition.labelForFormat || operator;
-            fn = (field, op, values, valueSrc, valueType, opDef, operatorOptions, isForDisplay) => {
-                return `${field} ${_operator} ${values}`;
-            };
+        if (!fn) {
+            const _operator = operatorDefinition.sqlOp || operator;
+            if (cardinality == 0) {
+                fn = (field, op, values, valueSrc, valueType, opDef, operatorOptions) => {
+                    return `${field} ${_operator}}`;
+                };
+            } else if (cardinality == 1) {
+                fn = (field, op, value, valueSrc, valueType, opDef, operatorOptions) => {
+                    return `${field} ${_operator} ${value}`;
+                };
+            } else if (cardinality == 2) {
+                fn = (field, op, values, valueSrc, valueType, opDef, operatorOptions) => {
+                    const valFrom = values.first();
+                    const valTo = values.get(1);
+                    return `${field} ${_operator} ${valFrom} AND ${valTo}`;
+                };
+            }
         }
         if (!fn)
             return undefined;
@@ -129,7 +172,7 @@ export const queryString = (item, config, isForDisplay = false) => {
         const fieldFullLabel = fieldPartsLabels ? fieldPartsLabels.join(config.settings.fieldSeparatorDisplay) : null;
         const fieldLabel2 = fieldDefinition.label2 || fieldFullLabel;
         const formatField = config.settings.formatField || defaultSettings.formatField;
-        const formattedField = formatField(fieldName, fieldParts, fieldLabel2, fieldDefinition, config, isForDisplay);
+        const formattedField = formatField(fieldName, fieldParts, fieldLabel2, fieldDefinition, config);
         
         //format expr
         const args = [
@@ -140,11 +183,10 @@ export const queryString = (item, config, isForDisplay = false) => {
             (valueTypes.length > 1 ? valueTypes : valueTypes[0]),
             omit(operatorDefinition, ['formatOp', 'mongoFormatOp', 'sqlFormatOp']),
             operatorOptions,
-            isForDisplay
         ];
         let ret = fn(...args);
         if (isRev) {
-            ret = config.settings.formatReverse(ret, operator, reversedOp, operatorDefinition, revOperatorDefinition, isForDisplay);
+            ret = config.settings.sqlFormatReverse(ret, operator, reversedOp, operatorDefinition, revOperatorDefinition);
         }
         return ret;
     }